+++ docpie/__init__.py
@@ -26,7 +26,7 @@ __all__ = ['docpie', 'Docpie',
 # it's not a good idea but it can avoid loop importing
 __version__ = Docpie._version
 
-__timestamp__ = 1563716272.669418  # last sumbit
+__timestamp__ = 1684215839.7577085  # last sumbit
 
 logger = getLogger('docpie')